+++ lib/i18n/backend/weeler.rb
@@ -9,6 +9,8 @@ require 'i18n/backend/weeler/importer'
 module I18n
   module Backend
     class Weeler
+      attr_accessor :i18n_cache
+
       PLURAL_KEYS = ["zero", "one", "other"]
 
       autoload :StoreProcs,  'i18n/backend/weeler/store_procs'
@@ -29,10 +31,51 @@ module I18n
           end
         end
 
+        def reload_cache
+          i18n_cache.clear
+
+          Translation.all.each do |translation|
+            i18n_cache.write [translation.locale, translation.key], translation if translation.value.present?
+          end
+
+          i18n_cache.write('UPDATED_AT', Settings.i18n_updated_at)
+        end
+
       protected
 
         def lookup(locale, key, scope = [], options = {})
           key = normalize_flat_keys(locale, key, scope, options[:separator])
+          return lookup_in_cache(locale, key, scope, options)
+        end
+
+      private
+
+        def lookup_in_cache locale, key, scope = [], options = {}
+          # reload cache if cache timestamp differs from last translations update
+          reload_cache if i18n_cache.read('UPDATED_AT') != Settings.i18n_updated_at
+
+          return nil if i18n_cache.read([:missing, [locale, key]])
+
+          keys = expand_keys key
+
+          keys.reverse.each do |check_key|
+
+            result = i18n_cache.read([locale, check_key])
+
+            return result.value unless result.blank?
+          end
+
+          # mark translation as missing
+          i18n_cache.write([:missing, [locale, key]], true)
+
+          if ::Weeler.create_missing_translations
+            return store_empty_translation locale, key, options
+          else
+            return nil
+          end
+        end
+
+        def lookup_in_database locale, key, scope = [], options = {}
           result = Translation.locale(locale).lookup(key).load
 
           if result.empty?
@@ -47,18 +90,16 @@ module I18n
               fallback_value = fallback_backend_translation locale, key
               translation.update_attributes value: fallback_value if fallback_value.present?
             end
-            translation.value
+            return translation.value
           else
             chop_range = (key.size + FLATTEN_SEPARATOR.size)..-1
             result = result.inject({}) do |hash, r|
               hash[r.key.slice(chop_range)] = r.value
               hash
             end
-            result.deep_symbolize_keys
+            return result.deep_symbolize_keys
           end
         end
-
-      private
 
         # For a key :'foo.bar.baz' return ['foo', 'foo.bar', 'foo.bar.baz']
         def expand_keys(key)
@@ -73,10 +114,16 @@ module I18n
           interpolations = options.keys - I18n::RESERVED_KEYS
           keys = options[:count] ? PLURAL_KEYS.map { |k| [key, k].join(FLATTEN_SEPARATOR) } : [key]
           keys.each do |key|
-            translation = Weeler::Translation.find_or_initialize_by locale: locale.to_s, key: key, interpolations: interpolations
+            translation = Weeler::Translation.find_or_initialize_by locale: locale.to_s, key: key
+            translation.interpolations = interpolations
             fallback_value = fallback_backend_translation locale, key
-            translation.value = fallback_value if fallback_value.present?
-            translation.save
+            if fallback_value.present?
+              translation.value = fallback_value
+              translation.save
+              reload_cache
+            else
+              translation.save
+            end
             return_value = translation.value
           end
           return_value