+++ lib/i18n/backend/weeler/importer.rb
@@ -27,6 +27,8 @@ module I18n
 
             end # sheets
 
+            I18n.backend.backends[0].reload_cache
+
           end # import
 
           private