+++ lib/weeler.rb
@@ -26,6 +26,12 @@ module Weeler
   mattr_accessor :mount_location_namespace
   @@mount_location_namespace = "weeler"
 
+  mattr_accessor :excluded_i18n_groups
+  @@excluded_i18n_groups = [:activerecord, :attributes, :helpers, :views, :i18n, :weeler]
+
+  mattr_accessor :i18n_cache
+  @@i18n_cache = ActiveSupport::Cache::MemoryStore.new
+
   def self.setup
     yield self
     if Weeler.use_weeler_i18n == true