+++ spec/dummy/db/migrate/20140123083707_translate_weeler_seos.rb
@@ -1,0 +1,15 @@
+class TranslateWeelerSeos < ActiveRecord::Migration
+  def self.up
+    Weeler::Seo.create_translation_table!({
+      title: :string,
+      description: :text,
+      keywords: :text
+    }, {
+      :migrate_data => true
+    })
+  end
+
+  def self.down
+    Weeler::Seo.drop_translation_table! :migrate_data => true
+  end
+end